+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This release contains **BREAKING** changes. Make sure to read and apply upgrade 
 
 - **[Breaking]** Remove ability to abort transactions using `throw(:abort)`. Please use `raise WaterDrop::Errors::AbortTransaction`.
 - **[Breaking]** Disallow (similar to ActiveRecord) exiting transactions with `return`, `break` or `throw`.
+- [Enhancement] Make variants fiber safe.
 
 ### Upgrade Notes